refactor: index.js 활용해서 import 부분 코드 줄이기
