+++ src/components/AD/index.js
@@ -1,0 +1,6 @@
+export { default as ADArticleModal } from "./ADArticleModal";
+export { default as ADCard } from "./ADCard";
+export { default as ADCardWrapper } from "./ADCardWrapper";
+export { default as ADCarousel } from "./ADCarousel";
+export { default as ADCategory } from "./ADCategory";
+export { default as ADDateTab } from "./ADDateTab";
