+++ src/components/Now/index.js
@@ -1,0 +1,3 @@
+export { default as FoodMenuCardCard } from "./FoodMenuCard";
+export { default as FoodMenuCarousel } from "./FoodMenuCarousel";
+export { default as HotKeyword } from "./HotKeyword";
